--- a/patrickstar/core/chunk_data.py
+++ b/patrickstar/core/chunk_data.py
@@ -31,6 +31,7 @@ import time
 import torch
 
 from patrickstar.core.memtracer import RuntimeMemTracer
+from patrickstar.manager.cuda_context import CUDAContext
 from patrickstar.profiler import profiler
 from patrickstar.utils import logger, getsizeof
 import patrickstar.utils.global_timer as global_timer
@@ -47,6 +48,7 @@ class Chunk(object):
         chunk_id: int,
         memory_tracer: RuntimeMemTracer,
         memory_cache: Optional[MemoryCache],
+        with_async_move: bool,
         local_rank: int = 0,
         is_dummy: bool = False,
     ):
@@ -89,6 +91,10 @@ class Chunk(object):
         self.with_mem_cache = memory_cache is not None
         if self.with_mem_cache:
             self.memory_cache = memory_cache
+
+        self.with_async_move = with_async_move
+        if self.with_async_move:
+            self.compute_finish_event = torch.cuda.Event()
 
     def is_dummy(self):
         return self._is_dummy
@@ -189,6 +195,13 @@ class Chunk(object):
         """
         self._state_dict[old_state] -= 1
         self._state_dict[new_state] += 1
+        if (
+            self.with_async_move
+            and old_state == TensorState.COMPUTE
+            and self._state_dict[TensorState.COMPUTE] == 0
+        ):
+            cuda_ctx = CUDAContext()
+            self.compute_finish_event.record(cuda_ctx.compute_stream)
 
     def get_state(self):
         """
@@ -242,6 +255,18 @@ class Chunk(object):
     def move(self, target_device: torch.device):
         r"""
         Move the chunk to `target_device`.
+        """
+        if self.with_async_move:
+            cuda_ctx = CUDAContext()
+            self.compute_finish_event.synchronize()
+            with torch.cuda.stream(cuda_ctx.copy_stream):
+                self.move_sync(target_device)
+        else:
+            self.move_sync(target_device)
+
+    def move_sync(self, target_device: torch.device):
+        r"""
+        Move the chunk to `target_device` synchronizely.
         NOTE() Please check if the `target_device` has enough room before.
 
         Args: