Tested on single card for model size 4B, 8B, 10B, 20B on V100 and A100.