@zhuzilin 2021-12-07
For GPT_DS_20B model, this will reduce from 30.2s to 26.0s.

Before:

```
LOSS of step 4: 36.15625
After step 4. using patrickstar, gradient checkpoint: True, fp16 True
MA 23886.34 MB         Max_MA 25422.34 MB         CA 29778.0 MB         Max_CA 36178 MB 
CPU Virtual Memory: used = 715.13 GB, percent = 71.0%
Step 4 elaspe 30.218841075897217 s, 88.74770329796964 Tflops
CLIENT_access_dist ........... 9.181479930877686, 30.384109135470922 %
CLIENT_release ............... 0.04853343963623047, 0.16061085334050834 %
CHUNK_LIST_prepare_device .... 3.9061248302459717, 12.926469810147594 %
chunk_cpu_gpu_move ........... 5.2403953075408936, 17.341947500408185 %
chunk_gpu_cpu_move ........... 5.702709197998047, 18.87187468080246 %
CHUNK_LIST_chunk_move ........ 5.7044970989227295, 18.877791349708552 %
CHUNK_LIST_make_room ......... 1.8412587642669678, 6.0932450608517446 %
FWD .......................... 6.089883089065552, 20.15314239027459 %
CLIENT_release_dist .......... 0.012448549270629883, 0.04119576391438546 %
BWD .......................... 12.072030544281006, 39.949757153050726 %
ADAM_prepare_data_grad_copy .. 2.3362889289855957, 7.7314396289820175 %
CLIENT_access ................ 0.02909398078918457, 0.09628019601839663 %
ADAM_prepare_data ............ 2.3743484020233154, 7.857389169919046 %
ADAM_compute ................. 5.522885322570801, 18.27678671053202 %
ADAM_param_fp32_to_fp16 ...... 3.9356894493103027, 13.024307481085398 %
ADAM_release_data ............ 0.03429770469665527, 0.11350078750318734 %
ADAM ......................... 12.056118726730347, 39.897100456674686 %
TOTAL ........................ 30.218032360076904
------------- DATA MOVE RESULTS --------------
chunk_cpu_gpu_move: 77568.0 MB, 101 times, 14801.93677152183 MB/s
chunk_gpu_cpu_move: 77568.0 MB, 101 times, 13601.956071551129 MB/s
ADAM_prepare_data_grad_copy: 39026.0937538147 MB, 409 times, 16704.309672331332 MB/s
ADAM_param_fp32_to_fp16: 78052.1875076294 MB, 409 times, 19831.896930106464 MB/s
******************** LOSS ********************
[1.125, 51.21875, 107.125, 74.75, 36.15625]
```

After:

```
LOSS of step 4: 36.15625
After step 4. using patrickstar, gradient checkpoint: True, fp16 True
MA 23886.34 MB         Max_MA 25422.34 MB         CA 37586.0 MB         Max_CA 37586 MB 
CPU Virtual Memory: used = 423.4 GB, percent = 42.1%
Step 4 elaspe 25.992978811264038 s, 103.17604462671454 Tflops
CLIENT_access_dist ........... 6.028136730194092, 23.192219398994055 %
CLIENT_release ............... 0.057527780532836914, 0.22132824243549878 %
CHUNK_LIST_prepare_device .... 2.492732286453247, 9.590358792097005 %
chunk_cpu_gpu_move ........... 3.496436357498169, 13.45193759649613 %
chunk_gpu_cpu_move ........... 3.183120012283325, 12.246506839904837 %
CHUNK_LIST_chunk_move ........ 3.1883277893066406, 12.266542866410475 %
CHUNK_LIST_make_room ......... 0.7393529415130615, 2.844533294504493 %
FWD .......................... 3.53817081451416, 13.612503742709203 %
CLIENT_release_dist .......... 0.01415705680847168, 0.05446684145393069 %
BWD .......................... 11.230754375457764, 43.20839608484503 %
ADAM_prepare_data_grad_copy .. 2.2543537616729736, 8.673238412424622 %
CLIENT_access ................ 0.02969050407409668, 0.11422910849120803 %
ADAM_prepare_data ............ 2.293386697769165, 8.823410921484589 %
ADAM_compute ................. 5.09198522567749, 19.590537477166055 %
ADAM_param_fp32_to_fp16 ...... 3.7332537174224854, 14.363051662861782 %
ADAM_release_data ............ 0.04077577590942383, 0.15687778552181664 %
ADAM ......................... 11.223139762878418, 43.179100172445764 %
TOTAL ........................ 25.992064952850342
------------- DATA MOVE RESULTS --------------
chunk_cpu_gpu_move: 77568.0 MB, 101 times, 22184.87398852665 MB/s
chunk_gpu_cpu_move: 77568.0 MB, 101 times, 24368.54397593344 MB/s
ADAM_prepare_data_grad_copy: 39026.0937538147 MB, 409 times, 17311.432844884614 MB/s
ADAM_param_fp32_to_fp16: 78052.1875076294 MB, 409 times, 20907.281801762518 MB/s
******************** LOSS ********************
[1.125, 51.21875, 107.125, 74.75, 36.15625]
```